# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OHLCVBar {
  /** ISO timestamp */
  date: string;
  /** @nullable */
  open?: number | null;
  /** @nullable */
  high?: number | null;
  /** @nullable */
  low?: number | null;
  /** @nullable */
  close?: number | null;
  /** @nullable */
  volume?: number | null;
}

export interface StockHistoryResult {
  symbol: string;
  period: string;
  interval: string;
  rowCount: number;
  data: OHLCVBar[];
  /** ISO timestamp when data was fetched */
  fetchedAt: string;
}

export interface StockInfo {
  symbol: string;
  /** @nullable */
  shortName?: string | null;
  /** @nullable */
  longName?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  exchange?: string | null;
  /** @nullable */
  regularMarketPrice?: number | null;
  /** @nullable */
  regularMarketChangePercent?: number | null;
  /** @nullable */
  marketCap?: number | null;
  /** @nullable */
  fiftyTwoWeekHigh?: number | null;
  /** @nullable */
  fiftyTwoWeekLow?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetStockHistoryParams = {
/**
 * Stock ticker symbol (e.g. AAPL, TSLA, 2330.TW)
 */
symbol: string;
/**
 * Time range - "1d" for today, "max" for full history
 */
period: GetStockHistoryPeriod;
/**
 * Data interval/resolution
 */
interval: GetStockHistoryInterval;
};

export type GetStockHistoryPeriod = typeof GetStockHistoryPeriod[keyof typeof GetStockHistoryPeriod];


export const GetStockHistoryPeriod = {
  '1d': '1d',
  max: 'max',
} as const;

export type GetStockHistoryInterval = typeof GetStockHistoryInterval[keyof typeof GetStockHistoryInterval];


export const GetStockHistoryInterval = {
  '1m': '1m',
  '5m': '5m',
  '15m': '15m',
  '1h': '1h',
  '1d': '1d',
} as const;

export type GetStockInfoParams = {
/**
 * Stock ticker symbol
 */
symbol: string;
};