# lib/api-zod/src/generated/types/oHLCVBar.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OHLCVBar {
  /** ISO timestamp */
  date: string;
  /** @nullable */
  open?: number | null;
  /** @nullable */
  high?: number | null;
  /** @nullable */
  low?: number | null;
  /** @nullable */
  close?: number | null;
  /** @nullable */
  volume?: number | null;
}