# lib/api-zod/src/generated/types/stockHistoryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OHLCVBar } from './oHLCVBar';

export interface StockHistoryResult {
  symbol: string;
  period: string;
  interval: string;
  rowCount: number;
  data: OHLCVBar[];
  /** ISO timestamp when data was fetched */
  fetchedAt: string;
}