# lib/api-zod/src/generated/types/stockInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StockInfo {
  symbol: string;
  /** @nullable */
  shortName?: string | null;
  /** @nullable */
  longName?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  exchange?: string | null;
  /** @nullable */
  regularMarketPrice?: number | null;
  /** @nullable */
  regularMarketChangePercent?: number | null;
  /** @nullable */
  marketCap?: number | null;
  /** @nullable */
  fiftyTwoWeekHigh?: number | null;
  /** @nullable */
  fiftyTwoWeekLow?: number | null;
}